+++ src/common/flac_crc.h
@@ -1,8 +1,10 @@
+#include <stdint.h>
+
 /*given a new byte and the previous checksum value,
   assigns a new checksum to that value*/
 
 void
-flac_crc8(int byte, void *checksum);
+flac_crc8(uint8_t byte, void *checksum);
 
 void
-flac_crc16(int byte, void *checksum);
+flac_crc16(uint8_t byte, void *checksum);
